Update Makefile tp be more complete

+++ Makefile
@@ -1,4 +1,6 @@
-compile:
+init: distclean tidy
+
+build:
 	go build -o build/ cmd/main.go
 
 run:
@@ -7,5 +9,21 @@ run:
 test:
 	go test ./...
 
+lint:
+	golangci-lint run ./...
+
 clean:
-	rm -r build
+	if test -d ./build; then rm -r build; fi
+
+distclean: clean
+	test -d "./vendor" && rm -r vendor
+	rm go.sum
+
+update: update_libs tidy
+
+update_libs:
+	go get -t -u ./...
+
+tidy:
+	go mod tidy
+	go mod vendor
